Migrate db to new indexes (#1784)
* Migrate db to new indexes

* first pass

* finish up

+++ apps/platform/pkg/adapt/postgresio/save.go
@@ -10,11 +10,11 @@ import (
 	"github.com/thecloudmasters/uesio/pkg/sess"
 )
 
-const INSERT_QUERY = "INSERT INTO public.data (id,uniquekey,collection,tenant,autonumber,fields) VALUES ($1,$2,$3,$4,$5,$6)"
-const UPDATE_QUERY = "UPDATE public.data SET uniquekey = $2, fields = fields || $4 WHERE id = $1 and collection = $3"
-const DELETE_QUERY = "DELETE FROM public.data WHERE id = ANY($1) and collection = $2"
-const TOKEN_DELETE_QUERY = "DELETE FROM public.tokens WHERE fullid = ANY($1)"
-const TOKEN_INSERT_QUERY = "INSERT INTO public.tokens (fullid,recordid,token,collection,tenant,readonly) VALUES ($1,$2,$3,$4,$5,$6)"
+const INSERT_QUERY = "INSERT INTO public.data (id,uniquekey,owner,createdby,updatedby,createdat,updatedat,collection,tenant,autonumber,fields) VALUES ($1,$2,$3,$4,$5,to_timestamp($6),to_timestamp($7),$8,$9,$10,$11)"
+const UPDATE_QUERY = "UPDATE public.data SET uniquekey = $2, owner = $3, createdby = $4, updatedby = $5, createdat = to_timestamp($6), updatedat = to_timestamp($7), fields = fields || $10 WHERE id = $1 and collection = $8 and tenant = $9"
+const DELETE_QUERY = "DELETE FROM public.data WHERE id = ANY($1) and collection = $2 and tenant = $3"
+const TOKEN_DELETE_QUERY = "DELETE FROM public.tokens WHERE recordid = ANY($1) and collection = $2 and tenant = $3"
+const TOKEN_INSERT_QUERY = "INSERT INTO public.tokens (recordid,token,collection,tenant,readonly) VALUES ($1,$2,$3,$4,$5)"
 
 func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 
@@ -26,26 +26,49 @@ func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 	readTokens := map[string][]string{}
 	resetTokenIDs := []string{}
 
-	collectionName, err := getDBCollectionName(request.Metadata, tenantID)
-	if err != nil {
-		return err
-	}
+	collectionName := request.Metadata.GetFullName()
 
 	batch := &pgx.Batch{}
 
-	err = request.LoopChanges(func(change *adapt.ChangeItem) error {
+	err := request.LoopChanges(func(change *adapt.ChangeItem) error {
 
 		fieldJSON, err := gojay.MarshalJSONObject(change)
 		if err != nil {
 			return err
 		}
-		fullRecordID := makeDBId(collectionName, change.IDValue)
-		uniqueID := makeDBId(collectionName, change.UniqueKey)
+
+		ownerID, err := change.GetOwnerID()
+		if err != nil {
+			return err
+		}
+
+		createdByID, err := change.GetCreatedByID()
+		if err != nil {
+			return err
+		}
+
+		updatedByID, err := change.GetUpdatedByID()
+		if err != nil {
+			return err
+		}
+
+		createdAt, err := change.GetFieldAsInt(adapt.CREATED_AT_FIELD)
+		if err != nil {
+			return err
+		}
+
+		updatedAt, err := change.GetFieldAsInt(adapt.UPDATED_AT_FIELD)
+		if err != nil {
+			return err
+		}
+
+		fullRecordID := change.IDValue
+		uniqueID := change.UniqueKey
 
 		if change.IsNew {
-			batch.Queue(INSERT_QUERY, fullRecordID, uniqueID, collectionName, tenantID, change.Autonumber, fieldJSON)
+			batch.Queue(INSERT_QUERY, fullRecordID, uniqueID, ownerID, createdByID, updatedByID, createdAt, updatedAt, collectionName, tenantID, change.Autonumber, fieldJSON)
 		} else {
-			batch.Queue(UPDATE_QUERY, fullRecordID, uniqueID, collectionName, fieldJSON)
+			batch.Queue(UPDATE_QUERY, fullRecordID, uniqueID, ownerID, createdByID, updatedByID, createdAt, updatedAt, collectionName, tenantID, fieldJSON)
 		}
 
 		if request.Metadata.IsWriteProtected() {
@@ -65,15 +88,13 @@ func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 	if deleteCount > 0 {
 		deleteIDs := make([]string, deleteCount)
 		for i, delete := range request.Deletes {
-			deleteIDs[i] = makeDBId(collectionName, delete.IDValue)
+			deleteIDs[i] = delete.IDValue
 		}
-		batch.Queue(DELETE_QUERY, deleteIDs, collectionName)
+		batch.Queue(DELETE_QUERY, deleteIDs, collectionName, tenantID)
 	}
 
-	collectionNameLength := len(collectionName) + 1
-
 	if len(resetTokenIDs) > 0 {
-		batch.Queue(TOKEN_DELETE_QUERY, resetTokenIDs)
+		batch.Queue(TOKEN_DELETE_QUERY, resetTokenIDs, collectionName, tenantID)
 	}
 
 	if len(readWriteTokens) > 0 {
@@ -82,7 +103,6 @@ func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 				batch.Queue(
 					TOKEN_INSERT_QUERY,
 					key,
-					key[collectionNameLength:],
 					token,
 					collectionName,
 					tenantID,
@@ -98,7 +118,6 @@ func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 				batch.Queue(
 					TOKEN_INSERT_QUERY,
 					key,
-					key[collectionNameLength:],
 					token,
 					collectionName,
 					tenantID,
